Updated as per comments
