+++ resources/conf/wdio-bstack-single.conf.js
@@ -19,11 +19,19 @@ var overrides = {
     browserName: 'Chrome',
     browser_version: "latest",
     acceptInsecureCerts: true,
-    name: (require('minimist')(process.argv.slice(2)))['bstack-session-name'] || (((require('minimist')(process.argv.slice(2)))['_'])[0].split('/').reverse())[0] +
-          " - " + new Date().toISOString(),
-    build: process.env.BROWSERSTACK_BUILD_NAME || 'browserstack-examples-webdriverio' + " - " + new Date().toISOString()
+    name: (require('minimist')(process.argv.slice(2)))['bstack-session-name'] || 'default_name',
+    build: process.env.BROWSERSTACK_BUILD_NAME || 'browserstack-examples-webdriverio' + " - " + new Date().getTime()
   }],
   afterTest: function (test, context, { error, result, duration, passed, retries }) {
+    console.log('name: ' + (require('minimist')(process.argv.slice(2)))['bstack-session-name'])
+    if((require('minimist')(process.argv.slice(2)))['bstack-session-name']) {
+      browser.executeScript("browserstack_executor: {\"action\": \"setSessionName\", \"arguments\": {\"name\":\"" +
+        (require('minimist')(process.argv.slice(2)))['bstack-session-name'] + " - " + new Date().getTime() + "\" }}");
+    } else {
+      browser.executeScript("browserstack_executor: {\"action\": \"setSessionName\", \"arguments\": {\"name\":\"" + test.title +
+        " - " + new Date().getTime() +  "\" }}");
+    }
+
     if(passed) {
       browser.executeScript('browserstack_executor: {"action": "setSessionStatus", "arguments": {"status":"passed","reason": "Assertions passed"}}');
     } else {
